pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fds-dev"
version = "0.1.0"
description = "A blazingly fast, structure-aware linter for your documentation, with built-in translation for non-native English speakers."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Flamehaven", email = "flamehaven01@users.noreply.github.com"}
]
keywords = [
    "linter",
    "documentation",
    "markdown",
    "translation",
    "i18n",
    "developer-tool",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/flamehaven01/FDS-Dev"
Repository = "https://github.com/flamehaven01/FDS-Dev"

[project.scripts]
fds = "fds_dev.main:cli"

[tool.setuptools]
packages = ["fds_dev"]
